Updated to new json schema
The new json schema is in annotations.schema.json. The change
in the schema is to add a rater to each line and modify the format
so that we can add additional pluera_lines and b_lines and
have a rater for each line per frame. Merging multiple json files
from different raters will be a lot easier with the new schema.

The rest of the schema remains the same, but the annotations now
look like:

  "frame_annotations": [
    {
      "pleura_lines": [
        {
          "rater": "andrew",
          "line": {
            "points": [
              [
                -123.78386340849477,
                -55.79519150558471,
                0.0
              ],
              [
                -133.903449327777,
                -59.25065986826647,
                0.0
              ]
            ]
          }
        },
        {
          "rater": "andrew",
          "line": {
            "points": [
              [
                -151.42761030994868,
                -51.84608480537702,
                0.0
              ],
              [
                -176.35634635500983,
                -53.08018064919193,
                0.0
              ]
            ]
          }
        }
      ],
      "b_lines": [
        {
          "rater": "andrew",
          "line": {
            "points": [
              [
                -142.54212023448136,
                -125.15137792798254,
                0.0
              ],
              [
                -169.44540962964635,
                -126.13865460303447,
                0.0
              ]
            ]
          }
        }
      ],
      "frame_number": 0
    } ...

Added conversion script to convert a directory with old jsons to
new format jsons as well as a validator that validates the new
jsons against the schema.

Also modified AnnotateUltrasound to allow the rater to enter their
name and prompt if not present when saving. Annotations are now
saved using the new format. This is work in progress - for now
we just need to agree on the schema changes.