Use free list for mrmlnodes (#142)
* Use free list for mrmlnodes

Make sure we free back to the free list and fetch from the free list
first before allocating a whole new mrml node so that we aren't always
allocating mrml nodes from Slicer. Also block nested callbacks, some
of which will actually come from AdjudicateUltrasound, but we need to
block in AnnotateUltrasound.

* Added useFreeList to choose whether to use free list or not, set to False

For now, I am disabling the free list as there is some subtle issue
with reusing a node that is causing annotations to break that I
haven't got to the bottom of. Leaving the free list code in for now
so I can keep debugging while freeing up memory correctly when not
using free list.

* Move many info logs to debug

* Skip sector drawing when we have a single point, unset and set observers only if already present

* Remember showHideLines across frames, but reset on new clip

* Refresh display after show/hide lines and overlay, overlay has to be redrawn if enabled

* Rater name entry should not require hitting enter

* Separate overlay drawing from depth guide drawing

* ClearAllLines should not go into infinite loop when rater has no lines

* Rater change should save to new file and update markups to new colors