AdjudicatorUltrasound extension subclassed from AnnotateUltrasound (#134)
* Added AdjudicatorUltrasound module

* Re-hookup click observer on Red view for line selection

* Handle both RedView clicks for selection and drawing

* Use a free list of mrml nodes for createMarkupLine

We maintain a free list, and try to allocate from it first and only
if there are no more free mrmlnodes in the free list do we allocate
more. They are freed back to the free list so that we can reuse them
if we need them again without allocating from Slicer.

* More management of free nodes, allocate and free

On loadNextSequence, we clear the scene, which means we also want
to free up all the markup nodes so that they can be reallocated
with the correct scene. Also make sure nodes aren't both in
pleuraLines and bLines and in freeMarkupNodes, that is just a
bug.

* Disable handling of selection node when doing annotation

Also removed duplicated code from AdjudicateUltrasound when
the code is unchanged from AnnotateUltrasound.

* Remove node from pleuraLines or bLines before putting into free list

* Disable free list

* Get rid of more duplicated code from Adjudicate and pull in fixes from Annotate

* UI changes for adjudicate

1. Fix next unlabeled to find next unadjudicated
2. Reordered the widgets to put input directory, read higher.
3. changed label to Adjudicator from Rater for Adjudicate.
4. Adjudicate and Annotate don't need to hit enter when changing rater name

* Fix the icon for Adjudicate to reuse Annotate

* Fix CMakeLists.txt for AdjudicateUltrasound

* Use both annotate and adjudicate pngs and py for adjudicate

* Update icon for AdjudicateUltrasound

* Delete incorrect calls to syncMarkupsToAnnotations after updating annotations

* Do not toggle overlay in show/hide lines

Overlay volume will not get drawn because lines aren't showing, but
depth guide should still show if it is enabled.

* Do not connect shortcuts in Adjudicate as Annotate already does it

Since Adjudicate inherits from Annotate, when Annotate creates
shortcuts, it is actually calling Adjudicate's shortcuts which
then call Annotate through super().

* Separate overlay drawing from depth guide drawing

* ClearAllLines should not go into infinite loop when rater has no lines

* Rater change should save to new file and update markups to new colors

* Disable drawing buttons and shortcuts for Adjudicator

Since adjudicator is not supposed to modify files, we are going
to disable drawing and adding/removing frames. Also pull in the
depth guide changes so we save the depth guide setting for
adjudicator.

* Adjust the size of the Overlay  [Space] button based on whether ClearAllLines is present