Fix leaking overlayVolume and depthGuideVolume (#146)
* Fix leaking overlayVolume and depthGuideVolume

We were just setting these to None without calling RemoveNode()
on them, causing the scene to keep a reference. Then we would
allocate a whole new volume for each in updateOverlayVolume
for depthGuideVolume and in loadNextSequence for overlayVolume,
so the memory would just keep leaking as we drew more lines,
went back and forth in frames etc.

Also made the line endpoints smaller to match Adjudicate, they
are too large currently for small screens.

* Also draw lines the same size as when updating them

* Added a test to cycle thru drawing lines and checking memory usage

* Also add playback to the memory test

* Change playback in test to 2.0fps