Select all, deselect all, copy, paste lines in Annotate (#149)
* Remove extraneous code and add missing freeList check

* Added Select All (cmd/ctrl-a), Copy (cmd/ctrl-c), Paste (cmd/ctrl-v)

Uses SelectAll, Copy, Paste and Escape key sequences that translate to
platform specific keyboard shortcuts (cmd/ctrl-a, cmd/ctrl-c, cmd/ctrl-v, esc)
etc. Will select all lines on a frame, copy all lines on a frame and then
paste to a new frame, but as owned by the rater and therefore rater colors.
Only for annotate.

* Added toggling selection on cmd+click (ctrl+click) on a line or control point

Adds or removes from the selected nodes list for eventual copy/paste.

* Add onPointRemove handler for pasted lines to fix deletion

Also added more user facing messages for select, unselect
and paste to tell the user how many lines were selected,
and pasted.

* Added a global key filter for CMD-A and Esc as they are special for Qt

* Also allow deletion of selected lines

* Limit Cmd-a/Cmd-click based selection options to Annotate only

We want to leave subclasses like Adjudicate to manage their own
global shortcuts.

* Use add and remove observers unconditionally

Basically Slicer is losing track or removing observers somewhere
and so just overrride addObserver and removeObserver to allow
adding and removing whether or not the observer is already added
or removed without throwing warnings (it already ignores, but the
warnings make things look bad). Change calls to add and remove
observers to just use a try/except block as all our nodes need
the same observers and so adding them unconditionally is ok when
we create and removing them in freeMarkupNode is also correct.

* Remove duplicate onDeleteSelectedLines, and make it user friendly

Ask the user if they are sure they want to delete lines if more than
one line is selected. Also clear the selection if lines are deleted
as the selections will not be valid after deletion.

* Add/Remove PointRemoved handler from logic only

It turns out that if you addObserver(..., self.logic.onPointRemoved) from
Widget, you can't removeObserver(..., self.onPointRemoved) from Logic as
the observers don't look the same even though they are calling the same
actual handler. So instead, add the onPointRemoved handler in
createMarkupLine by passing is whether to add it or not as an
argument which is only called from onPasteLines. Normally we only
add the handler after we have both points when adding lines.

* Set focus to redview forcefully

* Use a Custom Observer Mixin to avoid the warnings on add/remove observers

For some reason, still running into issues trying to use hasObserver before
adding or removing observers. We seem to miss removing some observers when
deleting all nodes and get errors. So now the mixing just allows us to
unconditionally add and remove observers (they are copies of the
VTKObservationMixin) but without the warnings. The warnings were just
warnings, they were not actually re-adding or re-removing anything and
could be ignored, but look ugly and fill up the python console.

* Added and fixed tests

Added force=False option to allow tests to send force=True for
onPasteLines and onDeleteSelectedLines to by pass duplicates
and dialog box in tests.