Added Cmd/Ctrl+X for Cut (#151)
* Added Cmd/Ctrl+X for Cut

Cut will copy all lines in the frame to clipboard, and cut only
the lines owned by the rater. It is basically Copy+Delete so
it copies all lines, but only Deletes what it is allowed to.
Pasting in a new frame will paste all the lines copied, not just
the ones that were owned by the rater.

Also cleaned up freeing of nodes from Clipboard to make sure
we don't leak inadvertantly.

* Do not ask for confirmation on Cut as that breaks flow

Cut requires multiple keys, so user is already trying to cut and
therefore we shouldn't query them with a dialog.